Last Merge

+++ resources/views/patients/show/show_diagnostic.blade.php
@@ -13,9 +13,7 @@
             <td>{!! implode(', ', (array)$item->diagnos_concreto) !!}</td>
             <td class="text-right">
                 @if ( in_array(Auth::user()->role_id,[1,3]) )
-                <a href="/deletediagnostic/{{$item->id}}" data-token="{{csrf_token()}}" class='btn btn-danger btn-xs'
-                    data-confirm="Seguro que quieres eliminar este pia?"
-                    onclick="return confirm('¿Estas segura que quieres el documento?')">
+                <a href="/deletediagnostic/{{$item->id}}" data-token="{{csrf_token()}}" class='btn btn-danger btn-xs' data-confirm="Seguro que quieres eliminar este pia?" onclick="return confirm('¿Estas segura que quieres el documento?')">
                     <i class="fas fa-trash"></i>
                 </a>
                 @endif
